experiment/stompclient/requirements.txt
stomp.py==6.0.0